scraper/requirements.txt
supabase>=2.3.0
google-generativeai>=0.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
curl_cffi>=0.7.1
tenacity>=8.3.0
